kun.py
I succeeded in getting this bot to work fine and hosted it through Repl.it. It was around the time when slash commands rolled in when I had to stop updating the bot source code and focus on my academics. I still haven't given up on making discord bots, but it has taken a back seat for now and I am no more adding new commands. However, I will be open to any suggestions or better ways to implement commands.